feature: jwt 토큰 기반 로그인 인증 구현

+++ user/src/main/java/com/nhnacademy/user/properties/JwtProperties.java
@@ -25,7 +25,7 @@ public class JwtProperties {
 
     private String secret;
 
-    private Integer expirationTime;
+    private Long expirationTime;
 
     private String tokenPrefix;
